файл main.js разит на модули
